Refactor admin dashboard: extract tab components, add pagination, fix reset bug
Split AdminDashboard.jsx into separate components under components/admin/:
TicketsTab, RevenueTab, TeamsTab, CoursesTab, UsersTab, Pagination, and
shared helpers. Add ticket grouping by paymentId, clickable rows, print
buttons, and pagination across all data tabs. Fix pagination not resetting
to page 1 on sort/filter changes by using useEffect instead of render-time
state mutation. Add auto-print support to TicketPage via ?print=true param.
